% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNN-10foldCV.R
\name{KNN_10CV}
\alias{KNN_10CV}
\title{Use K Near Neighbors method to make predictions}
\usage{
KNN_10CV(Xtrain, ytrain, Xtest, method)
}
\arguments{
\item{Xtrain}{The train data set without the response value, only the predictors}

\item{ytrain}{The response value in the train data set}

\item{Xtest}{The test data set}

\item{method}{The regression or classification}
}
\value{
The prediction results for the test data
}
\description{
This function can be used for both regression and classification problems
Using 10-fold cross validation to choose the best k for the model
}
\examples{
train<-titanic::titanic_train;
Xtrain<-train[,-c(2,4,11)];
ytrain<-train[,2];
test<-titanic::titanic_test;
test<-test[,-c(3,10)];
Xtrain=na.omit(Xtrain);
test$Age[is.na(test$Age)]=mean(test$Age,na.rm=T);
test$Fare[is.na(test$Fare)]=mean(test$Fare,na.rm=T);
KNN_10CV(Xtrain,ytrain,Xtest=test,method="regression")
}
